refactor(gridsat): reorganize PlatformMetadata class into its own module
